requirements.txt
fastapi>=0.111.0
fastapi[standard]
sqlmodel==0.0.14
pymysql==1.1.0
python-dotenv==1.0.0
pytest==8.0.0
httpx==0.26.0
scikit-learn==1.4.0
pytest